Added get route and made an axios request in agency page
