+++ backend/.env
@@ -1,0 +1,1 @@
+MONGO_URI=mongodb+srv://admin:upcmmgzqmrZfnCcy@dwa.1vehr.mongodb.net/DWA?retryWrites=true&w=majority
