+++ frontend/src/pages/Agency/Agency.js
@@ -1,15 +1,29 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect } from 'react';
 import CreateAgencyBtn from "../../components/CreateAgencyBtn/CreateAgencyBtn";
+import DataTable from "../../components/AgencyTable/DataTable";
+import axios from 'axios';
 
-class Agency extends Component {
-  render() {
-    return (
-      <div>
-        <h2>Agency</h2>
-        <CreateAgencyBtn />
+function AgencyTable() {
+  const [data, setData] = useState([]);
+  const [qFilter, setQFilter] = useState("");
+
+  useEffect(() => {
+    
+    axios.get('http://localhost:8000/agency/').then((res) => setData(res.data.data));
+
+  }, []);
+
+  console.log(data);
+  return (
+    <div className="agency-table">
+      <div id="filter"></div>
+      <div className="data-table-container">
+        <DataTable data={data} />
       </div>
-    );
-  }
+      <CreateAgencyBtn />
+    </div>
+  )
 }
 
-export default Agency;
+export default AgencyTable;
+
